requirements.txt
streamlit==1.29.0
pandas==2.0.3
numpy==1.24.4
datasets
scikit-learn==1.3.2
shap==0.43.0
matplotlib==3.7.3
seaborn==0.12.2
plotly
umap-learn
imbalanced-learn
altair
joblib==1.3.2
